Change resume cache to BG XML
Previously the data stored in redis was the processed candidate info. This would sometimes cause issues of not running any new parsing logic on parsed resumes. This would cause 'false positives' in the sense that the app could appear to be working fine but only because it kept responding with already parsed and processed data.