Error Messaging Overhaul and Integration
Per GET-1410 Jason and I have created new messaging for RPS to the client. These commits include the new error messages, custom error codes (new to RPS and will be in the Apiary after merge to develop), and the inclusion of those where appropriate. This new messaging covers 90%+ of messages sent to the client.